Debug the refactoring chores
