+++ src/components/logs/registeration-form.css
@@ -9,7 +9,7 @@ body {
 }
 
 .effect {
-  background-image: url("../assets/yacht-background-r.jpg");
+  background-image: url("../../assets/yacht-background-r.jpg");
   background-repeat: round;
   height: 100vh;
   filter: blur(5px);
